openclaw-thai-dashboard/requirements.txt
flask==3.1.0
flask-cors==5.0.1
requests==2.32.3
python-dotenv==1.1.0
gunicorn==23.0.0
cachetools==5.5.2
psutil==7.0.0
